Refactor transposition table
